Files: rename pocl_event to pocl_tracing
